% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.h.R
\name{anova}
\alias{anova}
\title{ANOVA}
\usage{
anova(data, dep = NULL, ind = NULL, modelTerms = NULL)
}
\arguments{
\item{data}{the data set as a data frame}

\item{dep}{a string specifying the dependent variable in \code{data}}

\item{ind}{a vector of strings specifying the independent variables in 
\code{data}}

\item{modelTerms}{a list of string vectors specifying the model terms}
}
\description{
ANOVA
}

