% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.h.R, R/jamovi.R, R/ttestis.h.R, R/ttestones.h.R, R/ttestps.h.R
\docType{data}
\name{anovaOptions}
\alias{anovaOptions}
\alias{jamovi}
\alias{ttestISOptions}
\alias{ttestOneSOptions}
\alias{ttestPSOptions}
\title{jamovi classes}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
anovaOptions

jamovi

ttestISOptions

ttestOneSOptions

ttestPSOptions
}
\description{
jamovi classes exported for jamovi to access (see https://www.jamovi.org)
}
\keyword{datasets}

