% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttestis.h.R
\name{ttestIS}
\alias{ttestIS}
\title{Independent Samples T-Test}
\usage{
ttestIS(data, dependent, grouping, alternative = "two.sided", mu = 0,
  varEqual = FALSE, confLevel = 0.95)
}
\arguments{
\item{data}{the data as a data frame}

\item{dependent}{a string specifying the dependent variable in \code{data}}

\item{grouping}{a string specifying the grouping variable in \code{data}; 
must have two levels}

\item{alternative}{\code{'two.sided'}, \code{'less'}, or \code{'greater'}}

\item{mu}{a number, the mean of the null hypothesis}

\item{varEqual}{TRUE or FALSE (default), assume equal variances}

\item{confLevel}{a number between 0 and 1 (default: 0.95), specifying the 
confidence interval width}
}
\description{
Independent Samples T-Test
}

