% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttestones.h.R
\name{ttestOneS}
\alias{ttestOneS}
\title{One Sample T-Test}
\usage{
ttestOneS(data, x, alternative = "two.sided", mu = 0, confLevel = 0.95)
}
\arguments{
\item{data}{the data as a data frame}

\item{x}{a string specifying the variable in \code{data}}

\item{alternative}{\code{'two.sided'}, \code{'less'}, or \code{'greater'}}

\item{mu}{a number specifying the mean of the null hypothesis}

\item{confLevel}{a number between 0 and 1 (default: 0.95) specifying the 
confidence interval width}
}
\description{
One Sample T-Test
}

