% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttestps.h.R
\name{ttestPS}
\alias{ttestPS}
\title{Paired Samples T-Test}
\usage{
ttestPS(data, x, y, alternative = "two.sided", mu = 0, varEqual = FALSE,
  confLevel = 0.95)
}
\arguments{
\item{data}{the data as a data frame}

\item{x}{a string specifying the first measurement in \code{data}}

\item{y}{a string specifying the second measurement in \code{data}}

\item{alternative}{\code{'two.sided'}, \code{'less'}, or \code{'greater'}}

\item{mu}{a number, the mean of the null hypothesis}

\item{varEqual}{TRUE or FALSE (default), assume equal variances}

\item{confLevel}{a number between 0 and 1 (default: 0.95), specifying the 
confidence interval width}
}
\description{
Paired Samples T-Test
}

